requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
argcomplete==3.5.1
bcrypt==4.1.2
blinker==1.9.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
cryptography==42.0.5
decorator==5.1.1
dnspython==2.6.1
f5-sdk-python==0.9.2
fastapi==0.115.8
Flask==3.1.0
gunicorn==23.0.0
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
paramiko==3.4.0
purestorage==1.19.0
py==1.11.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.6.3
PyNaCl==1.5.0
python-dotenv==1.0.1
pytz==2025.2
pyvmomi==8.0.2.0.1
requests==2.31.0
retry==0.9.2
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
sshtunnel==0.4.0
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.2.1
uvicorn==0.34.0
Werkzeug==3.1.3
